Shunting Yard Term Parsing Reuse
Shunting Yard now piggybacks off of the existing _nt_term function
to parse parenthesized portions of an expression and to handle function calls.
This concludes the features addition portion of the algorithm implementation.
Last thing to do is have it create a tree instead of a list in postfix order.