Removed Token Types XMacro
Up until now the string array that contained a string for each token type was generated
by an XMacro in conjunction with the enum itself during compile-time from the same values.
This is no longer the case due to the need to have descriprive strings for more generic token types.
As a result, the files tokens.h and tokens.c have been removed and the array remade and put into lexer.h.